+++ app/Http/Controllers/productController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\products\StoreProductRequest;
 use App\Http\Requests\products\UpdateProductRequest;
 use App\Models\Product;
+use App\Models\Role;
 use App\Models\User;
 use App\Notifications\ProductNotification;
 use App\Traits\HttpResponses;
@@ -23,6 +24,7 @@ class productController extends Controller
     use processImageTrait,HttpResponses,Products;
     public function index()
     {
+        abort_if(!$this->authorize('showAll_product'),403,'Unauthorized');
         $products = Product::with(['user:id,name', 'images:imageable_id,photo'])->get();
         return $this->success($products,'ok');
     }
@@ -45,6 +47,7 @@ class productController extends Controller
      */
     public function store(StoreProductRequest $request)
     {
+        abort_if(!$this->authorize('create_product'),403,'Unauthorized');
         $product =new Product();
         DB::transaction(function () use ($request, $product) {
            $product->fill($request->all())->save();
@@ -53,7 +56,8 @@ class productController extends Controller
               $product->images()->create(['photo' => $image])->save();
            if (!$product)
                 return $this->responseError('Product failed',404);
-           $admin =User::firstWhere('role','admin');
+           $admin =Role::firstWhere('name','Admin');
+           $admin= $admin->users()->first();
            $user =$product->user;
            $admin->notify(new ProductNotification($product->name,$user,null,'database'));
         });
@@ -69,6 +73,7 @@ class productController extends Controller
      */
     public function show(Product $product)
     {
+        abort_if(!$this->authorize('view_product'),403,'Unauthorized');
         return $this->success(
             $product->load('user:id,name','images:imageable_id,photo')
             ,'Product retrieved successfully');
@@ -83,6 +88,7 @@ class productController extends Controller
      */
     public function update(UpdateProductRequest $request,Product $product)
     {
+        abort_if(!$this->authorize('update_product'),403,'Unauthorized');
         DB::transaction(function () use ($request, $product) {
             $product->fill([$request->all()])->update();
             $oldImages = $product->images()->pluck('id')->toArray();
@@ -105,11 +111,14 @@ class productController extends Controller
      */
     public function destroy(Product $product)
     {
+        abort_if(!$this->authorize('delete_product'),403,'Unauthorized');
         DB::transaction(function () use ($product) {
-            foreach ($product->images as $pic) {
-                $this->deletePhoto($pic->photo);
+            if ($product->images->exists()) {
+                foreach ($product->images as $pic) {
+                    $this->deletePhoto($pic->photo);
+                }
+                $product->images()->delete();
             }
-            $product->images()->delete();
             $product->delete();
             if ($product)
                 return $this->responseError('deleted failed',404);
